+++ packages/client/src/services/API/UserAPI.ts
@@ -1,5 +1,5 @@
 import { Axios } from './Axios'
-import { API, PATHNAMES } from '../../utils/constants'
+import { API, PATHNAMES } from '@/utils'
 import { SignUpPayload } from './types'
 import { useNavigateCustom } from '@/hooks'
